Extract test helpers to a shared class
Created a shared TestHelpers class to eliminate duplicate code across test files.
- Moved runSdev, getOutput, and productsDirectory to TestHelpers
- Updated all test classes to use the shared helpers directly
